+++ src/spring/azext_spring/commands.py
@@ -36,7 +36,7 @@ def load_command_table(self, _):
 
     app_command = CliCommandType(
         operations_tmpl='azext_spring.app#{}',
-        client_factory=cf_spring_20220301preview
+        client_factory=cf_spring_20220501preview
     )
 
     app_managed_identity_command = CliCommandType(
@@ -144,7 +144,7 @@ def load_command_table(self, _):
                          table_transformer=transform_app_table_output)
         g.custom_show_command(
             'show', 'app_get', table_transformer=transform_app_table_output,
-            client_factory=cf_spring_20220301preview)
+            client_factory=cf_spring_20220501preview)
         g.custom_command('start', 'app_start', supports_no_wait=True)
         g.custom_command('stop', 'app_stop', supports_no_wait=True)
         g.custom_command('restart', 'app_restart', supports_no_wait=True)