+++ src/spring/azext_spring/app.py
@@ -52,13 +52,14 @@ def app_create(cmd, client, resource_group, service, name,
                enable_persistent_storage=None,
                persistent_storage=None,
                assign_endpoint=None,
-               loaded_public_certificate_file=None,
                enable_liveness_probe=None,
                enable_readiness_probe=None,
                enable_startup_probe=None,
                liveness_probe_config=None,
                readiness_probe_config=None,
-               startup_probe_config=None):
+               startup_probe_config=None,
+               assign_public_endpoint=None,
+               loaded_public_certificate_file=None):
     '''app_create
     Create app with an active deployment, deployment should be deployed with default banner
     1. Create app
@@ -85,6 +86,7 @@ def app_create(cmd, client, resource_group, service, name,
         'enable_persistent_storage': enable_persistent_storage,
         'persistent_storage': persistent_storage,
         'public': assign_endpoint,
+        'public_for_vnet': assign_public_endpoint,
         'loaded_public_certificate_file': loaded_public_certificate_file
     }
     create_deployment_kwargs = {
@@ -106,6 +108,7 @@ def app_create(cmd, client, resource_group, service, name,
     update_app_kwargs = {
         'enable_persistent_storage': enable_persistent_storage,
         'public': assign_endpoint,
+        'public_for_vnet': assign_public_endpoint
     }
 
     deployable = deployable_selector(**create_deployment_kwargs, **basic_kwargs)
@@ -140,6 +143,7 @@ def app_update(cmd, client, resource_group, service, name,
                deployment=None,  # set by validator
                # app
                assign_endpoint=None,
+               assign_public_endpoint=None,
                enable_persistent_storage=None,
                enable_ingress_to_app_tls=None,
                https_only=None,
@@ -196,13 +200,13 @@ def app_update(cmd, client, resource_group, service, name,
 
     app_kwargs = {
         'public': assign_endpoint,
+        'public_for_vnet': assign_public_endpoint,
         'enable_persistent_storage': enable_persistent_storage,
         'persistent_storage': persistent_storage,
         'loaded_public_certificate_file': loaded_public_certificate_file,
         'enable_end_to_end_tls': enable_ingress_to_app_tls,
         'https_only': https_only,
     }
-
     if deployment is None:
         updated_deployment_kwargs = {k: v for k, v in deployment_kwargs.items() if v}
         if updated_deployment_kwargs: