+++ src/spring/azext_spring/spring_instance.py
@@ -6,9 +6,7 @@
 # pylint: disable=wrong-import-order
 # pylint: disable=unused-argument, logging-format-interpolation, protected-access, wrong-import-order, too-many-lines
 from ._utils import (wait_till_end, _get_rg_location)
-from .vendored_sdks.appplatform.v2022_01_01_preview import models
 from .vendored_sdks.appplatform.v2022_05_01_preview import models
-from knack.log import get_logger
 from .custom import (_warn_enable_java_agent, _update_application_insights_asc_create)
 from ._build_service import _update_default_build_agent_pool
 from .buildpack_binding import create_default_buildpack_binding_for_application_insights
@@ -56,6 +54,7 @@ class DefaultSpringCloud:
                        reserved_cidr_range=None,
                        service_runtime_network_resource_group=None,
                        app_network_resource_group=None,
+                       enable_log_stream_public_endpoint=None,
                        zone_redundant=False,
                        sku=None,
                        tags=None,
@@ -65,6 +64,13 @@ class DefaultSpringCloud:
         properties = models.ClusterResourceProperties(
             zone_redundant=zone_redundant
         )
+
+        if enable_log_stream_public_endpoint is not None:
+            properties.vnet_addons = models.ServiceVNetAddons(
+                log_stream_public_endpoint=enable_log_stream_public_endpoint
+            )
+        else:
+            properties.vnet_addons = None
 
         if marketplace_plan_id:
             properties.marketplace_resource = models.MarketplaceResource(
@@ -146,6 +152,7 @@ def spring_create(cmd, client, resource_group, name,
                   gateway_instance_count=None,
                   enable_api_portal=False,
                   api_portal_instance_count=None,
+                  enable_log_stream_public_endpoint=None,
                   ingress_read_timeout=None,
                   marketplace_plan_id=None,
                   no_wait=False):
@@ -176,6 +183,7 @@ def spring_create(cmd, client, resource_group, name,
         'gateway_instance_count': gateway_instance_count,
         'enable_api_portal': enable_api_portal,
         'api_portal_instance_count': api_portal_instance_count,
+        'enable_log_stream_public_endpoint': enable_log_stream_public_endpoint,
         'marketplace_plan_id': marketplace_plan_id,
         'no_wait': no_wait
     }