+++ src/spring/HISTORY.md
@@ -2,8 +2,9 @@ Release History
 ===============
 1.1.0
 ---
-* New command `az spring create` has new argument "--ingress-read-timeout" to set ingress read timeout when create Azure Spring App.
-* New command `az spring update` has new argument "--ingress-read-timeout" to update ingress read timeout for Azure Spring App.
+* Command `az spring create` has new argument "--ingress-read-timeout" to set ingress read timeout when create Azure Spring Apps.
+* Command `az spring update` has new argument "--ingress-read-timeout" to update ingress read timeout for Azure Spring Apps.
+* Command `az spring create` has new argument "--marketplace-plan-id" to purchase SaaS resource against given plan for Azure Spring Apps.
 * Command `az spring app deploy` and `az spring app deployment create` has new argument "--build-cpu" and "--build-memory" to set cpu and memory during build process.
 * Commands `az spring app create`, `az spring app update`, `az spring app deploy`, `spring app deployment create`
 and `spring app deployment update` have new arguments "enable_liveness_probe", "enable_readiness_probe", "enable_startup_probe", "liveness_probe_config", "readiness_probe_config", "startup_probe_config" to customize the probe settings of user applications